+++ src/dynamic-server/finger-memory-web/src/test/java/work/nobility/fingermemoryweb/mapper/LexiconMapperTests.java
@@ -32,12 +32,12 @@ public class LexiconMapperTests {
 
     lexiconMapper.insertOneLexicon(lexicon1);
 
-    Lexicon lexicon2 = lexiconMapper.selectLexiconById((int) lexicon1.getLexiconId());
+    Lexicon lexicon2 = lexiconMapper.selectLexiconById(lexicon1.getLexiconId());
     lexicon2.setLexiconCreateDate(null);
 
     assertEquals(lexicon1, lexicon2);
-    lexiconMapper.deleteLexiconById((int) lexicon1.getLexiconId());
-    userMapper.deleteUserById((int) user.getUserId());
+    lexiconMapper.deleteLexiconById(lexicon1.getLexiconId());
+    userMapper.deleteUserById(user.getUserId());
   }
 
   @Test
@@ -59,11 +59,11 @@ public class LexiconMapperTests {
     lexicon1.setLexiconBaidescription("词库描述update");
     lexiconMapper.updateLexicon(lexicon1);
 
-    Lexicon lexicon2 = lexiconMapper.selectLexiconById((int) lexicon1.getLexiconId());
+    Lexicon lexicon2 = lexiconMapper.selectLexiconById(lexicon1.getLexiconId());
     lexicon2.setLexiconCreateDate(null);
 
     assertEquals(lexicon1, lexicon2);
-    lexiconMapper.deleteLexiconById((int) lexicon1.getLexiconId());
-    userMapper.deleteUserById((int) user.getUserId());
+    lexiconMapper.deleteLexiconById(lexicon1.getLexiconId());
+    userMapper.deleteUserById(user.getUserId());
   }
 }
