+++ src/dynamic-server/finger-memory-web/src/test/java/work/nobility/fingermemoryweb/mapper/WordCategoryMapperTests.java
@@ -19,28 +19,28 @@ public class WordCategoryMapperTests {
 
     wordCategoryMapper.insertOneWordCategory(wordCategory1);
 
-    WordCategory wordCategory2 = wordCategoryMapper.selectWordCategoryById((int) wordCategory1.getCategoryId());
+    WordCategory wordCategory2 = wordCategoryMapper.selectWordCategoryById(wordCategory1.getCategoryId());
 
     assertEquals(wordCategory1, wordCategory2);
 
-    wordCategoryMapper.deleteWordCategoryById((int) wordCategory1.getCategoryId());
+    wordCategoryMapper.deleteWordCategoryById(wordCategory1.getCategoryId());
   }
 
   @Test
   public void updateWordCategoryTest() {
     WordCategory wordCategory1 = new WordCategory();
     wordCategory1.setCategoryName("分类名");
-    wordCategoryMapper.deleteWordCategoryById((int) wordCategory1.getCategoryId());
+    wordCategoryMapper.deleteWordCategoryById(wordCategory1.getCategoryId());
 
     wordCategoryMapper.insertOneWordCategory(wordCategory1);
 
     wordCategory1.setCategoryName("分类名update");
     wordCategoryMapper.updateWordCategory(wordCategory1);
 
-    WordCategory wordCategory2 = wordCategoryMapper.selectWordCategoryById((int) wordCategory1.getCategoryId());
+    WordCategory wordCategory2 = wordCategoryMapper.selectWordCategoryById(wordCategory1.getCategoryId());
 
     assertEquals(wordCategory1, wordCategory2);
 
-    wordCategoryMapper.deleteWordCategoryById((int) wordCategory1.getCategoryId());
+    wordCategoryMapper.deleteWordCategoryById(wordCategory1.getCategoryId());
   }
 }
